detect/http.host: rename file for consistency
